requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
tavily-python>=0.3.0
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0